requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
distlib==0.3.8
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langsmith==0.1.67
MarkupSafe==2.1.5
marshmallow==3.21.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.3
packaging==23.2
platformdirs==4.2.2
pyarrow==15.0.0
pydantic==2.7.2
pydantic_core==2.18.3
pylance==0.11.1
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
SQLAlchemy==2.0.30
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
torch==2.3.0
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.2
yarl==1.9.4
